requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
psycopg2-binary
python-dotenv
ffmpeg-python
